Skip reachability checks and upload pack when performing a fetch
While performing a fetch with a virtual file system we know that there will be missing objects and we don't want to download them just because of the reachability of the commits.  We also don't want to download a pack file with commits, trees, and blobs since these will be downloaded on demand.

This flag will skip the first connectivity check and by returning zero will skip the upload pack. It will also skip the second connectivity check but continue to update the branches to the latest commit ids