Lower the default slop value and remove check for object file when traversing uninteresting commits
The slop value is used to get a number of uninteresting
commit when traversing the commit graph.  When running
a status we only want the counts of interesting
commits so we don't need to look at the uninteresting
ones.  We also don't need to check for the object
files when marking parents as uninteresting.