PR 153303: Don't show local changes after doing an optimized "checkout -b foo"
Don't show local changes on the optimized "checkout -b" path as the optimized path does not compute the set of changes so can't report them correctly.  Since this path doesn't make any changes (other than creating a new reference and pointing head to it) the results of a "git status" before and after will be identical.

Related work items: #682832