Merged PR 164508: fix pager atexit logic to work correctly with performance atexit logic
I had to move the pager atexit registration to before the performance atexit registration to fix a bug where the performance traces got lost when the pager ran (due to the pager closing stdout/stderr).  As a result of moving this first, I was able to simplify the logic a little in pager.c because the function is guaranteed to be registered as it is now the first line of code that executes.

This is just fixing a bug in upstream git that we should submit and reduce the delta between our branch and upstream.