pre-reset hook: fix compile error
The pre-reset hook was called via run_hook_ve() which wants to get a
va_list, i.e. a vararg. However, we have an argv_array at hand, and
there is no easy or portable way to construct a va_list from that. Let's
just use the run_hook_argv() function that we just introduced.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>