Fix reset when using the sparse-checkout feature.
When using the sparse checkout feature the git reset command will add entries to the index that will have the skip-worktree bit off but will leave the working directory empty.  File data is lost because the index version of the files has been changed but there is nothing that is in the working directory.  This will cause the next status call to show either deleted for files modified or deleting or nothing for files added.  The added files should be shown as untracked and modified files should be shown as modified.

To fix this when the reset is running if there is not a file in the working directory and if it will be missing with the new index entry or was not missing in the previous version, we create the previous index version of the file in the working directory so that status will report correctly and the files will be availble for the user to deal with.

Signed-off-by: Kevin Willford <kewillf@microsoft.com>