Merged PR 182673: Add support for read-object as a background process to retrieve missing objects
This PR converts the existing read_object hook proc model for downloading missing blobs to use a background process that is started the first time git encounters a missing blob and stays running until git exits.  Git and the read-object process communicate via stdin/stdout and a versioned, capability negotiated interface as documented in Documentation/technical/read-object-protocol.txt.  The advantage of this over the previous hook proc is that it saves the overhead of spawning a new hook process for every missing blob.

The model for the background process was refactored from the recent git LFS work.  I refactored that code into a shared module (sub-process.c/h) and then updated convert.c to consume the new library.  I then used the same sub-process module when implementing the read-object background process.