Add check for the end of the entries for the thread partition. Add test for lazy init name hash with specific directory structure
The lazy init hash name was causing a buffer overflow when the last entry in the index was multiple folder deep with parent folders that did not have any files in them.

This adds a test for the boundary condition of the thread partitions with the folder structure that was triggering the buffer overflow.

The fix was to check if it is the last entry for the thread partition in the handle_range_dir and not try to use the next entry in the cache.

Signed-off-by: Kevin Willford <kewillf@microsoft.com>