squash! rebase: turn on progress option by default for format-patch
rebase: turn on progress option by default for format-patch

This change passes the progress option of format-patch when run in
an interactive terminal, unless rebase runs in quiet mode.

The idea is to give the user an early warning if they try to rebase an
insane amount of commits by mistake.

Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Kevin Willford <kewillf@microsoft.com>