Merged PR 223027: unpack-trees: preserve index extensions
Make git checkout (and other unpack_tree operations) preserve the untracked cache status. This is valuable as it speeds up subsequent git operations that _would_ have used the untracked cache but previously had to regenerated it.

Ultimately, the untracked cache is caching information about the state of the working directory.  The fact that we create a new index from a different commit doesn't change the state of the working directory.  As a result, that doesn't impact the validity of the cache so copying it over is perfectly valid.  Later operations may update the working directory but 1) don't change the untracked files (by definition) and 2) are already well tested and handled by the untracked cache logic.

I ran all existing git tests with the patch applied and they all passed.  This only really tells us that it didn't break anything because untracked cache is turned off by default but it does show us that it still passes the untracked cache specific test cases (as they obviously turn it on).

I then modified the test_create_repo() function in test-lib-functions.sh to turn on the untracked cache feature after creating the test repo and ran all the tests again twice - the first time without the patch and again with the patch.  This run
is more interesting because it is testing that having the untracked cache turned (with and without the patch) on doesn't break anything.

There were two test scripts that had failures with the patch:

t7063-status-untracked-cache.sh failed the test "not ok 1 - core.untrackedCache is unset"  This is actually a positive result because it is showing that I successfully turned on the untracked cache feature.

t1700-split-index.sh failed several tests in both runs (with and without patch) and upon examining the tests and their failures they are to be expected and do not indicate any bug.  Specifically, the failures were caused because the
tests check the sha of the index against a hard coded value in the test script. Because the untracked cache is turned on, the sha of the index does not match that hard coded value.  I edited several of the tests to update the sha they are checking against to match the sha actually generated and the tests pass.