Synchronize the sub-process API to upstream's `master`
Upstream Git saw the contribution of Ben's original implementation of
the sub-process API.

Unsurprisingly, that version did not survive the Git mailing list
unchanged, but Ben worked hard and eventually got a final version
integrated.

Let's update to that final version.

The update of convert.c is easy: simply use the upstream version. The
update of the read object hook simply imitates the diff in convert.c.