Include config.h in gvfs
The topic branch that was integrated into Git for Windows master to fix
the "aliases in a worktree" bug introduces a whole slur of
config-related changes, including a new header that every user of config
functions must include. This re-design made it already in upstream Git's
master branch, so we will have to adapt to it eventually, anyway.